+++ pmj_server/src/gui_main.rs
@@ -1,8 +1,20 @@
 use iced;
 
-mod gamemodes;
+use pmj_server_lib::{self, gamemodes};
+use pmj_shared;
 
 fn main() -> iced::Result {
-    //TODO
-    gamemodes::gui_base::main()
+    let config = pmj_server_lib::shared::read_server_config();
+    match config.gamemode {
+        pmj_shared::shared::GameModes::Base => {
+            gamemodes::gui_base::main()?;
+        }
+        pmj_shared::shared::GameModes::V1Simple => {
+            println!("還未支援！");
+        }
+        pmj_shared::shared::GameModes::V2Better => {
+            println!("還未支援！");
+        }
+    }
+    Ok(())
 }
