+++ pmj_shared/src/gamemodes_shared/mod.rs
@@ -19,7 +19,4 @@
 pub mod shared_base;
 
 // v1 - `simple`
-pub mod sharedv1_simple;
-
-// v2 - `better`
-pub mod sharedv2_better;
+//pub mod sharedv1_simple;
